Upgrade to Inkscape 1.1-dev (2020-05-11)
Use the correct path to invoke the "inkscape" program in the snap
package (/snap/bin/inkscape). See:

Snap fails on Ubuntu 16.04 with "libgsl.so.23: cannot open shared object
file" <https://gitlab.com/inkscape/inbox/-/issues/2721>

Note that Inkscape 1.0 sometimes converts the text to paths even without
the "--export-text-to-path" option, and it seems to do so randomly. That
problem has been fixed in Inkscape 1.1. In fact, the only significant
difference in the SVG files between Inkscape versions 0.92.4 and 1.1 is
that version 1.1 rounds the font size to four decimal places instead of
five.